Cache hosted whitelists for 5 minutes by default
When using Azure Blob Storage for hosting whitelists, headers get reset
every time the whitelist is saved and must be manually re-added to
prevent the whitelist from being re-downloaded for every request.

So as a convenience, default to 5 minute caching unless explicitly
prohibited or overridden by the response headers. This is allowed by
HTTP specs.